+++ packages/conta_bold_design_system/lib/conta_bold_design_system.dart
@@ -41,6 +41,7 @@ export 'src/bold_palette.dart';
 export 'src/bold_pontos_de_pagina.dart';
 export 'src/bold_vinho.dart';
 export 'src/bold_fonts.dart';
+export 'src/bold_type.dart';
 export 'src/bold_fundamentos.dart';
 export 'src/bold_gradients.dart';